pyproject.toml
[tool.poetry]
name = "recipe-app"
version = "0.1.0"
description = "Recipe App"
authors = ["akumars1 <Abhishek@deepfence.io>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.6"
Django = "^4.2.2"
drf-spectacular = "^0.26.2"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
